Fix basic_command_line test to allow for different pipe paths
The `basic_command_line` test would start the `riak attach` test
by searching for the string `erlang.pipe.1` on the first line.
In some cases you will use `erlang.pipe.X` where X is an increasing
integer.  This change simply removes that requirement from the
top of the test and checks for the proceeding string instead.