Don't call lager in remote functions
It is not safe to compile code, that calls lager, with Riak Test and
then run that code on the remote nodes.  Depending on the version the
remote node will have a different lager version.  If the API has
changed between versions then the calls will fail and leave a
non-obvious error message for the test runner to decipher.  This is
exactly what happened to me recently as the upcoming 1.3 release is
following lager master but Riak Test is using lager 1.2.1.  There is
an api change between those two causing the info call to fail.

https://github.com/basho/lager/compare/1.2.1...d6482649#L5L54