Add backup restore test
This is a port of systest-backup-and-restore from basho_expect with some
changes and additions. Biggest change was not using search-cmd to index
data for search, as it does not seem to update docs, just indexes. Then
added restore on a clean cluster. Basically it does this:

* Writes data, including a search enabled bucket
* Verifies written data, including searches
* Runs backup
* Deletes, modifies some data
* Verifies mods and deletes took
* Runs restore
* Verifies deleted data is back, mods stay (last write wins)
* Stops cluster and wipes out all data dirs
* Verifies clean cluster
* Runs restore
* Verifies data was restored, including searches

Also, adding some more noise to some riak test setup functions. This was
useful when initially debugging the test before it worked the first
time.