Avoid spurious ownership handoffs
While looking at issue #145 noticed that ownership handoffs were
happening after build_cluster returned. This in combination with a bug
in pipes + handoffs that Bryan is fixing broke the link walk urls test.
Trying to avoid this by checking for pending changes after ownership is
shared by all and agreed on. The test used to get into this state every
several runs. I've run it dozens of times with this and seems to help.