Retry on listkeys timeout
loaded upgrade has a tendency to fail on listkeys time out, which should
not make the fail test. Made the PB client have the auto_reconnect flag
set and added functions to retry reconnecting (as it doesn't happen
instantaneously) and retry the list keys a few times with exponentially
bigger time outs.