Add default node config
To be used by the Ubuntu 32 bit tester to reduce the ring size and allow
tests to pass on that limited memory environment. But it will take any
default node configuration you put in the rt_default_config parameter.
These are defaults: they are applied before any configuration given in
the code to *all* nodes.  Configuration passed by the test overrides it.
Example:
   { rt_default_config,
        [ {riak_core, [ {ring_creation_size, 16} ]} ] }

This addresses issue #176