Move the timing of the runnable test check to earlier in test run
Check if potential test modules are compatatible with the riak_test
API as part of the determination of which tests to run in the
riak_test:main function instead of doing the check in the
riak_test_runner module. This speeds up execution of test runs by
avoiding the penalty of attempting to stop all nodes for modules that
are not actually runnable tests.