Fix sometime failure fetching counter after partition
Depending on which partitions returned an answer first, fetching
a counter after the partition may return 'not_found'. Since we
expect there to at least one value on each side of the partition
get the counter with r=1 and notfound_ok=false so that the get fsm
will always wait for the vnode with data to answer.