Add wait until really deleted function
I was having test failures due to the delete operation timing out
silently. This was due to dependencies not update by rebar in the end.
But in any case, this new function is more robust in detecting when
tombstones are really reaped instead of a sleep call.
Also, matching delete operation so it's noisy when it fails.