Fix race checking logs for busy dist port
The test detects the busy dist port event, and immediately fetches the
logs and parses them looking for it in the output. But it's failing
because the message appears an instant after the logs are fetched.
This changes it so it will retry a few times as we do with most things.