Add option to allow tests in a directory to be explicitly skipped
Add skip option for use with the dir option to allow specified tests
in a directory to be skipped during a test run. This is useful in
cases where a directory contains a large number of tests of which a
small number are only appropriate to run in certain circumstances. A
good example of such a case is providing a way for open source users
to build and run the RiakCS tests and avoid the failure that would
occur trying to run the replication tests.