Remove checks for keys and buckets during transfers from verify_listkeys
There is way to ensure what portion of the full harvest is returned by
a coverage request during handoffs. A node may start accepting
requests before the data has been handed off and there is no way to
prevent or detect this and no way to have a deterministic test
condition that checks for the presence of all expected keys. This
change removes any checks during transfers. It also and adds a
directive to wait until transfers have completed after a node is added
or removed and before listkeys result verification resumes.