test that riak uses the correct search system for "search" mr inputs
Riak attempts to choose the correct search system (riak_search or
yokozuna) to provide MapReduce inputs when {search, Bucket, Query} is
specified. The `mapred_search_switch` test makes sure that the expected
system is chosen, given the cluster's configuration (whether riak_search
and yokozuna are each enabled, and whether the choice of which to use
has been set explicitly).