Fix test log formatting.
riak_test_lager_backend was supporting Lager 2.0-style messages by
manually extracting and then reconstituting the pieces. It's much
easier just to pass it to lager_default_formatter, and then the format
is more declarative. Since we also have the opportunity to inject
newlines in each line as it is produced, we can skip the fold that
joined the log lines and simply convert it to a binary with the
unicode module.