extend sink ack timeout, for slow test machines
It seems like 10ms might be too short for some of our builders, so we can
get an additional timeout trace message, which breaks the log matching
later in this test. One second should be long enough for any VM, right?

The other option is to expect any number of log messages, and just find
the one we need. I'd like to avoid that for now, just in case the
intermittent failure we've seen is *not* a simple additional timeout.