Remove any specific 1.2.1 code.
Unfortunately, this drastically reduces the value of the test because is
no longer verfies a "liveness" property of capabilities; it doesn't
verify that things will negotiate to the newer version.  However,
leaving the remainder of this test in place verifies two things: a.)
"safety", things do not regress to a earlier capability, and that the
override capabilities work as designed.