Fix partition repair stuck on vnode cmd
Changing from infinity (that's too long) to configure max waiting time.
This does not resolve other issues with the test that are still being
investigated: lots of data not found on first riak_kv vnode repaired in
some testers. But does prevent it from completely blocking riak_test
which happened on many machines.