Fix verify_backup_restore.
First, convert all of the node operations that write/delete to use the
Riak PB connection opened at the start of the test.  When deleting,
ensure we wait for the object to be fully deleted.

Here's where it gets more interesting:

In the original allow_mult=false behaviour, original objects are
replaced with modified objects, and then when performing the restore
operation, the originals are abandoned for the modified objects because
they have a greater timestamp.

In the new allow_mult=true behaviour, original objects are replaced with
modified objects, and then the restore operations restores the original
object as a *sibling*, because the objects have divergent vector clocks.

Adapt the test for this.