Test for 2i AAE repair
This test uses intercepts to fake a disconnect between the index data in
the object and what's stored on disk. Then it runs the new 2i repair
console command to verify that it's been fixed.

This copies a utility from repl_util to wait on AAE tree builds into
rt.erl. It should be removed from repl later.

To make the test work in the riak/1.4 branch, I had to use 1.4 branches
of the erlang clients.