init ~/rt/riak even if releases aren't built
This change allows initialization of ~/rt/riak to proceed even if
releases aren't built, or if you run this script outside of your
test-releases directory. It will report this to the user however.

This is useful if you only want to run riak tests against current. New
users of riak test and people working on bugs on crazy branches will
especially want this since they likely do not want to take the time to
build multiple releases that may not even be relevant to the bug.