Test 2i queries in mixed cluster
The test is far from perfect. It can't use the HTTP client on the old
nodes because it's incompatible. But the PB client seems fine.
To test all possible mixed cluster scenarios, you would have to manually
change your riak_test configuration to point to different versions. It
would be nice to have a better way to run the same test in a number of
combinations.