Add tree rebuild in 2i AAE test
This helps verify the problem where 2i AAE trees got dropped during a
rebuild, breaking future runs of the repair-2i command.
Enhanced the wait for AAE trees function to try a lock on the built
trees. The building process may have notified its end, but it is still
possible to get a not_built if you try to get a lock very quickly. This
attempts the lock for each partition before declaring total victory.