Improve tests.
Improve test infrastructure by extracing out common code into functions
that can be used to build out further test cases.

Label our existing unidirectional test as a "simple" test, and begin
building out a more exhaustive test case.

The exhaustive test case is not complete and currently a *work in
progress*. However, it currently configures three clusters and sets up
replication from one source to two sinks.  This is as far as the test
gets because this currently *fails* in setting up the initial
connections to the two clusters causing fullsync to completely abort.

It appears, but it unconfirmed, that a crash is causing cluster_conn to
fail and lose information about how to connect to the clusters
documented due to async storage of the remote cluster information to the
ring.  (I believe this to be the issue as this bug was discovered by
Kelly and I during debugging of another repl related race condition
during startup.)  Removing the second cluster connection call, and
letting it proceed with a connection to only one cluster allows the test
to proceed.