Add AAE test
This test verifies that AAE repairs replicas of values without passive
read repairs. This includes missing replicas and replicas with divergent
values.  It will also repair entire KV partitions lost, and if
configured for trees to rebuild, it will recover from AAE data loss and
corruption.

This version differs from the original 1.4 test only in the handling of
siblings. It does get before put for modifications and merges values by
choosing the longest one, as modifications in this test append bits.