Change heartbeat timeout to seconds in repl_rt_heartbeat test
The heartbeat timeout enforcement was recently updated to be
specified in seconds to match the documentation for that option. The
repl_rt_heartbeat test has since been failing since it still specified
the timeout in milliseconds.  This change makes the test use seconds
for the heartbeat timeout gets the test passing again.