Fix client_java_verify failures
The client_java_verify test has failed for quite some time with a
series of java.lang.NoClassDefFoundError errors. There seems to be an
issue with the use or generation of the jar file that includes all
needed dependencies for the test to run properly. Change the test
execution to use maven as is prescribed in the java client
documentation and add a dependency on the java client repository to
riak_test.