Add ability to handle identical siblings in rt:systest_read
Add optional parameter to rt:systest_read that provides the ability to
compare siblings and treat those with identical values and metadata
excluding the dot as a single value for the purposes of testing.

This is to facilitate testing with DVV enabled since there are several
known cases where use of DVV leads to siblings created internally by
Riak. One example of such a case is a write during handoff that is
forwarded by the vnode, but also executed locally.

The existing behavior of rt:systest_read is to fail when any siblings
are encountered with a badmatch. This behavior is maintained as the
default so existing tests maintain their current semantics without
explicit change.