Made test more reliable to check for leaks.
rpc or other processes will cause one or two persistent processes to start up during
the test, but the number shoould not increase after that. Altered the test to check for
massive increase in processes over a larger number of faults rather than an equality
of process_count over a relatively small number of faults.