Persuade rtdev harness to set advanced.config variables
In the case that no advanced.config file exists (everycase!) rt
would not add any advanced config settings to the conf.

This PR teaches rtdev to create an advanced.comfig file if none exists
so that tests may set advanced config.

In this case we set ring_size and also the `crdt_mixed_versions` app env