Restructured how connectivity tests are configured and run [RIAK-1608].
Due to leftover configuration attributes from prior tests within this module (see RIAK-1607), the order of tests has to be managed.  In order to accomplish this and be able to see exactly what configurations are used for each test, configuration blocks have been more directly associated with each connectivity test.