Added a configuartion sleep before first jmx dump
If the jmx dump is done before the jvm is up and running, the test fails because
the connection is refused. This can be replicated manually by starting a cluster
and attempting the same jmx dump command after the cluster has started, but
before the jvm is up.

This is intended to resolve an occasinal failure of a build run.