First EQC test for secondary indexing
It needs more love to match the functionality of the regular 2i tests,
but this is a good beginning. It can show a race in 2i pagination that
has been in the code for years. The race does not show up in run in a
single node with very few vnodes. I tried 32.  Ideally we could use
PULSE to make scheduling in the riak nodes less deterministic and help
find this kind of issue, but currently PULSE is a single node joint.

This test uses a different bucket for each EQC command run. It is not
perfect isolation, but it is good enough and very fast compared to
taking down the entire cluster in between each run.

(cherry picked from commit fa493858373642b4219273fcae3f98cd34f1c7d1)