introduce wrkrnd for sudden rpc timeout from SM; with code reshuffle and fixes
There are now two tests:

* bdp_sm_startstop, which runs the start/stop cycle on and from
  various cobinations of the 3-node cluster;

* bdp_spark, which only starts and stops a single service just now,
  but eventually will run the scenario "Primary Master to Standby
  Master failover" (I.A), with spark log grepping etc.

The workaround is, specifically, for the case when SM returns {error,
timeout}, but eventually succeeds after some 10 sec of trying,
apparently due to riak_ensemble not being at all times available.
This is a bug, to be fixed in data_platform_global_state:call/3 or
thereabout.