Fleshed out but disabled http client support
The http client seems to have a lot of problems and works differently
than the PB client.  Index values are not coming back on a get, even
though a PB get returns them, so indexing is not working properly.
Binaries are not encoded properly when the query index URL is created,
and empty binaries are taken to mean "no value".