Induce race condition in riak_snmp_stat_poller
There is a race in riak_snmp_stat_poller, which occurs when realtime
replication is enabled for more than two clusters. If the replication is
enabled quickly enough, then the bug does not manifest itself. This
change introduces a sleep between enabling replication between the
individual clusters, thus causing the test to fail consistently instead
of sporadically.