a test to check TS records put to eleveldb are not reversed
This is a series of 2 commits from branch
feature-mw-add-eleveldb-order-test, squashed/cherry-picked to current
riak_ts manually because Micah has started his branch from a point way
too old.

--- This is the 1st commit message:

First pass test, doesn't yet check order.

This is just testing getting the intercepter in place, and filtering
down to a single vnode.

--- This is the 2nd commit message:

Test actually verifies if orders match.

Order submitted by client are checked against the order of the puts
submitted to a vnode's eleveldb backend.

Author:    Micah Warren <mwarren@basho.com>
Date:      Mon Oct 26 16:24:29 2015 -0600

With a fix to a call to timeseries_util:create_bucket/2, which has since
become /3.